+++ tests/test_libfv.py
@@ -675,10 +675,12 @@ def test_libfv_get_version_number_string(allfonts):
 
 def test_libfv_get_version_number_string_bad_version_number():
     fv = FontVersion("tests/testfiles/Test-VersionOnly.ttf")
-    # mock a bad version number substring
-    fv.set_version_number("x.xxx")
-    response = fv.get_version_number_string()
-    assert len(response) == 0
+
+    with pytest.raises(ValueError):
+        # mock a bad version number substring
+        fv.set_version_number("x.xxx")
+
+    assert fv.get_version_number_string() == ""
 
 
 def test_libfv_get_version_number_tuple():
@@ -706,10 +708,11 @@ def test_libfv_get_version_number_tuple_bad_version_number():
     fv = FontVersion("tests/testfiles/Test-VersionOnly.ttf")
     assert fv.get_version_number_tuple() == (1, 0, 1, 0)
 
-    # mock a bad version number substring
-    fv.set_version_number("x.xxx")
-    response = fv.get_version_number_tuple()
-    assert response is None
+    with pytest.raises(ValueError):
+        # mock a bad version number substring
+        fv.set_version_number("x.xxx")
+
+    assert fv.get_version_number_tuple() is None
 
 
 def test_libfv_get_version_string_method():
@@ -919,6 +922,17 @@ def test_libfv_set_version_number(allfonts):
     assert prelength == postlength
     assert fv.version_string_parts[0] == "Version 2.000"
     assert fv.version == "Version 2.000"
+    assert fv.head_fontRevision == 2.000
+
+
+def test_libfv_set_version_number_invalid_number(allfonts):
+    fv = FontVersion(allfonts)
+
+    with pytest.raises(ValueError):
+        # mock a bad version number substring
+        fv.set_version_number("x.xxx")
+        response = fv.get_version_number_string()
+        assert len(response) == 0
 
 
 def test_libfv_set_version_string_one_substring():
@@ -926,11 +940,15 @@ def test_libfv_set_version_string_one_substring():
     fv.set_version_string("Version 2.000")
     assert len(fv.version_string_parts) == 1
     assert fv.version_string_parts[0] == "Version 2.000"
+    assert fv.version == "Version 2.000"
+    assert fv.head_fontRevision == 2.000
 
     fv2 = FontVersion("tests/testfiles/Test-VersionOnly.otf")
     fv2.set_version_string("Version 2.000")
     assert len(fv2.version_string_parts) == 1
     assert fv2.version_string_parts[0] == "Version 2.000"
+    assert fv2.version == "Version 2.000"
+    assert fv2.head_fontRevision == 2.000
 
 
 def test_libfv_set_version_string_two_substrings():
@@ -939,12 +957,16 @@ def test_libfv_set_version_string_two_substrings():
     assert len(fv.version_string_parts) == 2
     assert fv.version_string_parts[0] == "Version 2.000"
     assert fv.version_string_parts[1] == "DEV"
+    assert fv.version == "Version 2.000"
+    assert fv.head_fontRevision == 2.000
 
     fv2 = FontVersion("tests/testfiles/Test-VersionOnly.otf")
     fv2.set_version_string("Version 2.000;DEV")
     assert len(fv2.version_string_parts) == 2
     assert fv2.version_string_parts[0] == "Version 2.000"
     assert fv2.version_string_parts[1] == "DEV"
+    assert fv.version == "Version 2.000"
+    assert fv.head_fontRevision == 2.000
 
 
 def test_libfv_set_version_string_three_substrings():
@@ -954,6 +976,8 @@ def test_libfv_set_version_string_three_substrings():
     assert fv.version_string_parts[0] == "Version 2.000"
     assert fv.version_string_parts[1] == "DEV"
     assert fv.version_string_parts[2] == "other stuff"
+    assert fv.version == "Version 2.000"
+    assert fv.head_fontRevision == 2.000
 
     fv2 = FontVersion("tests/testfiles/Test-VersionOnly.otf")
     fv2.set_version_string("Version 2.000;DEV;other stuff")
@@ -961,6 +985,8 @@ def test_libfv_set_version_string_three_substrings():
     assert fv2.version_string_parts[0] == "Version 2.000"
     assert fv2.version_string_parts[1] == "DEV"
     assert fv2.version_string_parts[2] == "other stuff"
+    assert fv2.version == "Version 2.000"
+    assert fv2.head_fontRevision == 2.000
 
 
 def test_libfv_write_version_string_method(allfonts):
@@ -970,12 +996,16 @@ def test_libfv_write_version_string_method(allfonts):
     fv.write_version_string(fontpath=temp_out_file_path)
     fv2 = FontVersion(temp_out_file_path)
     assert fv2.version_string_parts[0] == "Version 2.000"
+    assert fv2.version == "Version 2.000"
+    assert fv2.head_fontRevision == 2.000
     # modify again to test write to same temp file path without use of the fontpath parameter in
     # order to test the block of code where that is handled
     fv2.set_version_number("3.000")
     fv2.write_version_string()
     fv3 = FontVersion(temp_out_file_path)
     assert fv3.version_string_parts[0] == "Version 3.000"
+    assert fv3.version == "Version 3.000"
+    assert fv3.head_fontRevision == 3.000
 
     os.remove(temp_out_file_path)
 
@@ -988,12 +1018,16 @@ def test_libfv_write_version_string_method_ttfont_object(allfonts):
     fv.write_version_string(fontpath=temp_out_file_path)
     fv2 = FontVersion(temp_out_file_path)
     assert fv2.version_string_parts[0] == "Version 2.000"
+    assert fv2.version == "Version 2.000"
+    assert fv2.head_fontRevision == 2.000
     # modify again to test write to same temp file path without use of the fontpath parameter in
     # order to test the block of code where that is handled
     fv2.set_version_number("3.000")
     fv2.write_version_string()
     fv3 = FontVersion(temp_out_file_path)
     assert fv3.version_string_parts[0] == "Version 3.000"
+    assert fv3.version == "Version 3.000"
+    assert fv3.head_fontRevision == 3.000
 
     os.remove(temp_out_file_path)